Begun to extract funtion into separate files.
